+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- DevHub series directive.
+
 ## [v0.3.2] - 2020-02-24
 
 ### Fixed